+++ lib/src/reload_report.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2015, the Dart project authors.  Please see the AUTHORS file
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
@@ -8,7 +8,7 @@ VMReloadReport newVMReloadReport(Map json) {
   return new VMReloadReport._(json);
 }
 
-/// The version of the VM service protocol (*not* the Dart VM itself).
+/// The reload report that is returned by `VMIsolateRef.reloadSources`.
 class VMReloadReport {
   /// Did the reload succeed or fail?
   final bool status;